delete tour helpers and rename files
